a better implementation for handler
